ath79: add support for AVM Fritzbox 3390 WASP SoC
This adds support for the Fritzbox 3390 WASP SoC (Wireless Assistant
Support Processor, sometimes referred to as Scorpion).
This is a memory only target on the mentioned Lantiq devices, that
has no access to leds, NAND, GPIO, USB on the Lantiq host.
The wifi adapters with 2.4GHz is connected to this SoC and is not
accessible from the Lantiq host. The ethernet port is connected to
the switch on the Lantiq host and allows network communication.
The Lantiq can load a network boot loader to this SoC using gpio.
The WASP SoC runs fully from RAM and is booted by using a remoteproc
kernel module and runs its own OpenWrt instance.

Specifications:
 - SoC: AR9342 on Lantiq host
 - RAM: 64 MB
 - Wireless:
   · Qualcomm-QCA9381 w/ 3×3 MIMO for 2.4GHz 802.11b/g/n

Installation:
For booting WASP, build the ATH79 target x7490 WASP image, copy it
to /lib/firmware on the Lantiq host image and use the network boot
firmware from the AVM firmware together with the avm_wasp kernel
module.

Signed-off-by: Daniel Kestrel <kestrel1974@t-online.de>